+++ src/module/apps/RollDialog.ts
@@ -22,6 +22,12 @@ export default class RollDialog extends FormApplication<
       template: 'systems/swade/templates/apps/rollDialog.hbs',
       classes: ['swade', 'roll-dialog'],
       width: 400,
+      filters: [
+        {
+          inputSelector: 'input.searchbox',
+          contentSelector: '.selections',
+        },
+      ],
       height: 'auto' as const,
       closeOnSubmit: true,
       submitOnClose: false,
@@ -63,9 +69,16 @@ export default class RollDialog extends FormApplication<
       this._addModifier();
       this.render();
     });
-    html.find('button.add-preset').on('click', () => {
-      this._addPreset();
+    html.find('.modifier .add-preset').on('click', (ev) => {
+      this._addPreset(ev);
       this.render();
+    });
+    html.find('button.toggle-list').on('click', (ev) => {
+      const target = ev.currentTarget as HTMLButtonElement;
+      const width = getComputedStyle(target).width;
+      html.find('.fas.fa-caret-right').toggleClass('rotate');
+      html.find('.searchbox').outerWidth(width, true);
+      html.find('.dropdown').outerWidth(width).slideToggle({ duration: 200 });
     });
     html.find('button[type="submit"]').on('click', (ev) => {
       this.extraButtonUsed = ev.currentTarget.dataset.type === 'extra';
@@ -79,10 +92,11 @@ export default class RollDialog extends FormApplication<
     });
   }
 
-  async getData(): Promise<object> {
+  async getData() {
     const data = {
-      rollModes: CONFIG.Dice.rollModes,
+      baseDice: this.ctx.roll.formula,
       displayExtraButton: true,
+      rollModes: CONFIG.Dice.rollModes,
       modGroups: CONFIG.SWADE.prototypeRollGroups,
       extraButtonLabel: '',
       rollMode: game.settings.get('core', 'rollMode'),
@@ -106,7 +120,7 @@ export default class RollDialog extends FormApplication<
     return data;
   }
 
-  protected async _updateObject(ev: Event, formData: FormData) {
+  protected override async _updateObject(ev: Event, formData: FormData) {
     const expanded = foundry.utils.expandObject(formData) as RollDialogFormData;
     Object.values(expanded.modifiers ?? []).forEach(
       (v, i) => (this.ctx.mods[i].ignore = v.ignore),
@@ -183,8 +197,6 @@ export default class RollDialog extends FormApplication<
       }
     }
 
-    this._markWilDie(terms);
-
     //recreate the roll
     const finalizedRoll = Roll.fromTerms(terms, roll.options);
 
@@ -203,6 +215,21 @@ export default class RollDialog extends FormApplication<
     return finalizedRoll;
   }
 
+  protected override _onSearchFilter(
+    event: KeyboardEvent,
+    query: string,
+    rgx: RegExp,
+    html: HTMLElement,
+  ) {
+    for (const li of Array.from(html.children) as HTMLLIElement[]) {
+      if (li.classList.contains('group-header')) continue;
+      const btn = li.querySelector('button');
+      const name = btn?.textContent;
+      const match = rgx.test(SearchFilter.cleanQuery(name!));
+      li.style.display = match ? 'block' : 'none';
+    }
+  }
+
   private _buildRollForEvaluation() {
     return Roll.fromTerms([
       ...this.ctx.roll.terms,
@@ -216,31 +243,9 @@ export default class RollDialog extends FormApplication<
     ]);
   }
 
-  /**
-   * This is a workaround to add the DSN Wild Die until the bug which resets the options object is resolved
-   * @param terms Array of roll terms
-   */
-  private _markWilDie(terms: RollTerm[]): void {
-    if (!game.dice3d) return;
-    for (const term of terms) {
-      if (term instanceof PoolTerm) {
-        for (const roll of term.rolls) {
-          for (const term of roll.terms) {
-            if (term instanceof WildDie) {
-              const colorPreset =
-                game.user?.getFlag('swade', 'dsnWildDie') ?? 'none';
-              if (colorPreset !== 'none') {
-                setProperty(term.options, 'colorset', colorPreset);
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
   /** add a + if no +/- is present in the situational mod */
   private _sanitizeModifierInput(modifier: string): string {
+    if (modifier.startsWith('@')) return modifier;
     if (!modifier[0].match(/[+-]/)) return '+' + modifier;
     return modifier;
   }
@@ -296,25 +301,21 @@ export default class RollDialog extends FormApplication<
     }
   }
 
-  private _addPreset() {
-    const select =
-      this.form!.querySelector<HTMLSelectElement>('#preset-selection')!;
-    const option = select.options[select.selectedIndex];
-    const index = Number(option.dataset.index);
+  private _addPreset(ev: JQuery.ClickEvent) {
+    const target = ev.currentTarget as HTMLButtonElement;
     const group = CONFIG.SWADE.prototypeRollGroups.find(
-      (v) => v.name === option.dataset.group,
+      (v) => v.name === target.dataset.group,
     );
-    if (!group) return;
-    const modifier = group.modifiers[index];
-    if (!modifier) return;
-    this.ctx.mods.push({
-      label: modifier.label,
-      value: modifier.value,
-    });
+    const modifier = group?.modifiers[Number(target.dataset.index)];
+    if (modifier) {
+      this.ctx.mods.push({
+        label: modifier.label,
+        value: modifier.value,
+      });
+    }
   }
 
-  /** @override */
-  close(options?: Application.CloseOptions): Promise<void> {
+  override close(options?: Application.CloseOptions): Promise<void> {
     //fallback if the roll has not yet been resolved
     if (!this.isResolved) this.resolve(null);
     $(document).off('keydown.chooseDefault');